requirements.txt
streamlit==1.53.1
anthropic==0.77.0
mistralai==1.12.3
pdf2image==1.17.0
pypdfium2==5.5.0
python-docx==1.2.0
fpdf2==2.8.6
python-dotenv==1.2.1
pillow==12.1.0
msal==1.35.0
